Added ob_start() comment
I have a 15 year old site with 500 posts and lots of legacy hard mapping. I also restructured my folder structure with folder names optimized for seo and moved all images to their own subdomain. This resulted in significant 404 errors.

Previously the site was on wordpress. Now its built with Astro on nodejs. Wordpress allows you to enter slugs from anywhere and the slugs will resolve to the correct post or page. Now that all posts are in a folder, I need to ensure that all slugs can be entered directly after the domain name and resolve to the proper folder. 

My .htaccess file reroutes all 404 errors to this router script. The script checks for known routing issues and provides the proper route or if no proper route found defaults to the 404 page. This gives me time to go through 500 posts and clean up all the legacy hard links and other issues.

The site is almost entirely static I wanted to use the nodejs routing but that would require me switching to new hosting and paying more. Instead, I will use php for dynamic features, i.e. routing, form responses, members access, etc. 

I use self-hosted mongodb for managing the static website content but will use MySQL or MariaDB for user registration because it is free with my hosting plan. This file will be updated for that purpose and may hook in earlier than the 404 trigger.