Simple JSON-LD Component in Astro
This is a simple Astro component I built to place a JSON-LD script on each of my pages. The page passes in title, description, and canonical url for each page using Astro.props and customizes the JSON with that data. 

If you aren't familiar with the props functionality, it can be confusing, especially when TypeScript issues are added into the explanations. 

It took me a while to figure out how to do this cleanly in Astro having had NO React experience and being new to javascript and nodejs but I think this is the cleanest way to do it. This could work just as easily in a javascript component. The Props are passed when you call the component in a one to one configuration. It doesn't matter what other chains of data may be involved to get the data into the page you are calling the component from. I struggled with thinking that I had to worry about the entire chain of data exchange between pages, layout pages, and multiple components.

This is an example of calling the component from within the content section of the Astro page, <PageLd {title} {canonical} {description} />. All Astro is doing is creating a custom html element with the name PageLd which is actually a variable I created when I imported the component at the top of the Astro page, "import PageLd from '@components/json-ld/PageLd.astro';"

the "@components" portion of the path is defined in in the tsconfig.json file and it looks like this:
"paths": {
	"~/assets/*": ["src/assets/*"],
	"@components/*": ["src/components/*"],
	"@layouts/*": ["src/layouts/*"],
},

In the component I'm calling, props is passed as an array that you simply destructure as needed like this: 
const { title, description, canonical } = Astro.props;

You don't have to destructure all the variables in props, just the ones you need. As long as you use the same variable names in the calling page and in the component, you only need the variable name. This keeps things very simple. 

With the data loaded in the component, I simply create a text string using template literals. The variables are passed into the template literal with ${variable} syntax. 

Then in the body of the component, you enter a script tag and you use the template directive "set:html={stringVariable}. If you don't use the set:html directive, the script won't be passed to the final html file properly. 

I find that I prefer using the template literals and javascript rather than using Astro's jsx functionality down in the body section. There are fewer restrictions on what you can and cannot do and its just less frustrating when you aren't intimately familiar with all the limitations of Astro jsx.

Although I create all the pages in our site, you can easily add variables for author name, image, and url if you need more. 

I like simple basic examples so that I can learn the bare bones of making something work. Then I like to layer new skills on top of that as needed. This way you can quickly get something up and running and then add complexity as you learn.