Simple PHP router hooking 404 and correcting known routing errors
This is a simple routing file that hooks into 404 using .htaccess and processes known routing issues before serving 404 page. I have around 500 posts from over 15 years and moved off of Wordpress onto Astro. There is a lot of hardcoded legacy content that I am working to fix but that takes time. 

The other thing this does is ensures that when any post slug is entered just after the domain name, the url will be rerouted to the blog folder. I pull a text file from my collection of posts whenever new posts are added and load an array from the text file to do this. 

I have a different text file and array for page issues rather than post issues. This one has to be created manually because all pages are created in Astro as Astro files. They are not stored in a database but will be at some point. 

This is a super simple routing file and should be helpful for anyone who is new to routing and just wants to get their feet wet with something that actually works before going deeper.